Converted ticker to enum in C++ and Python templates

+++ exchange/src/common/types/ticker.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <cstddef>
-
+#include <cstdint>
 #include <optional>
 #include <stdexcept>
 #include <string>
@@ -9,7 +8,8 @@
 #include <utility>
 
 namespace nutc::common {
-enum class Ticker : std::size_t { ETH = 0, BTC = 1, LTC = 2 }; // NOLINT
+// NOTE: this must be the same as Side in template.hpp
+enum class Ticker : std::uint8_t { ETH = 0, BTC = 1, LTC = 2 }; // NOLINT
 static constexpr auto TICKERS = {Ticker::ETH, Ticker::BTC, Ticker::LTC};
 
 inline std::string
